backend/requirements-updated.txt
fastapi==0.125.0
langchain==1.2.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.2
langchain-text-splitters==1.1.0
markdown-it-py==4.0.0
openai==2.13.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
python-multipart==0.0.21
qdrant-client==1.16.2
tiktoken==0.12.0
unstructured==0.18.21
unstructured-client==0.42.6
uvicorn==0.38.0
